Split src/registry/refresh.test.ts (warren-a0e8 / pl-9088 step 12)
Extract shared fixtures (buildSpawn, showOk, renderMissingParent,
rendered, CFG, FAKE_CLONE) into src/registry/refresh.test-helpers.ts
and split the 1041-line refresh.test.ts into four focused suites:

- refresh.test.ts (214 lines) - refreshAgentRegistry tests
- refresh.project.test.ts (240 lines) - basic refreshProjectAgents tests
- refresh.cross-tier.test.ts (306 lines) - cross-tier inheritance (warren-44a3)
- refresh.cache.test.ts (249 lines) - on-disk rendered cache (warren-44e3)

Each file is now well under the 500-line per-file / per-function budget.
buildSpawn was decomposed into handleRender + handleShow + a
respondFromHandler helper so its cognitive complexity drops below the
biome ceiling without the previous override.

- Remove biome noExcessiveLinesPerFunction override for refresh.test.ts
- Remove refresh.test.ts from noExcessiveCognitiveComplexity override
- Remove refresh.test.ts from scripts/file-size-budgets.json
- All 26 pre-existing tests pass; bun run check:all is green.
