+++ src/registry/refresh.test.ts
@@ -1,132 +1,9 @@
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
-import { mkdtemp, readdir, readFile, rm } from "node:fs/promises";
-import { tmpdir } from "node:os";
-import { join } from "node:path";
 import { openDatabase, type WarrenDb } from "../db/client.ts";
 import { AgentsRepo } from "../db/repos/agents.ts";
 import { DrizzleAdapter } from "../db/repos/drizzle-adapter.ts";
-import { ProjectsRepo } from "../db/repos/projects.ts";
-import { CanopyClient, type SpawnFn, type SpawnResult } from "./canopy.ts";
-import type { CanopyRegistryConfig } from "./config.ts";
-import {
-	RENDERED_CACHE_SUBPATH,
-	type RenderedCacheWriter,
-	refreshAgentRegistry,
-	refreshProjectAgents,
-} from "./refresh.ts";
-import type { AgentDefinition } from "./schema.ts";
-
-const CFG: CanopyRegistryConfig = {
-	repoUrl: "https://example.com/agents.git",
-	localDir: "/tmp/canopy-refresh",
-	cnBinary: "cn",
-	gitBinary: "git",
-};
-
-function buildSpawn(
-	listResp: unknown,
-	renderResponses: Record<string, { ok?: unknown; exit?: number; stderr?: string }>,
-	showResponses: Record<string, { ok?: unknown; exit?: number; stderr?: string }> = {},
-): SpawnFn {
-	return async (cmd) => {
-		if (cmd[1] === "list") {
-			return { stdout: JSON.stringify(listResp), stderr: "", exitCode: 0 };
-		}
-		if (cmd[1] === "render") {
-			const name = cmd[2] as string;
-			const handler = renderResponses[name];
-			if (!handler) {
-				return { stdout: "", stderr: `unhandled render: ${name}`, exitCode: 2 };
-			}
-			if (handler.exit !== undefined && handler.exit !== 0) {
-				return {
-					stdout: handler.ok !== undefined ? JSON.stringify(handler.ok) : "",
-					stderr: handler.stderr ?? "",
-					exitCode: handler.exit,
-				};
-			}
-			return { stdout: JSON.stringify(handler.ok), stderr: "", exitCode: 0 };
-		}
-		if (cmd[1] === "show") {
-			const name = cmd[2] as string;
-			const handler = showResponses[name];
-			if (!handler) {
-				// Default: structured "Prompt not found" envelope so the compose
-				// resolver treats unknown names as absent at this tier.
-				return {
-					stdout: JSON.stringify({
-						success: false,
-						command: "show",
-						error: `Prompt '${name}' not found`,
-					}),
-					stderr: "",
-					exitCode: 1,
-				};
-			}
-			if (handler.exit !== undefined && handler.exit !== 0) {
-				return {
-					stdout: handler.ok !== undefined ? JSON.stringify(handler.ok) : "",
-					stderr: handler.stderr ?? "",
-					exitCode: handler.exit,
-				};
-			}
-			return { stdout: JSON.stringify(handler.ok), stderr: "", exitCode: 0 };
-		}
-		const result: SpawnResult = { stdout: "", stderr: "unexpected cmd", exitCode: 1 };
-		return result;
-	};
-}
-
-/** Helper: build a structured `cn show --json` success envelope. */
-function showOk(
-	name: string,
-	body: {
-		version?: number;
-		sections?: Record<string, string>;
-		extends?: string;
-		mixins?: string[];
-		frontmatter?: Record<string, unknown>;
-	},
-) {
-	return {
-		ok: {
-			success: true,
-			command: "show",
-			prompt: {
-				id: `canopy-${name}`,
-				name,
-				version: body.version ?? 1,
-				sections: Object.entries(body.sections ?? {}).map(([n, b]) => ({ name: n, body: b })),
-				...(body.extends !== undefined ? { extends: body.extends } : {}),
-				...(body.mixins !== undefined ? { mixins: body.mixins } : {}),
-				...(body.frontmatter !== undefined ? { frontmatter: body.frontmatter } : {}),
-				status: "active",
-				createdAt: "2026-01-01T00:00:00.000Z",
-				updatedAt: "2026-01-01T00:00:00.000Z",
-			},
-		},
-	};
-}
-
-/** Helper: structured "Prompt 'X' not found" render error envelope. */
-function renderMissingParent(parentName: string) {
-	return {
-		exit: 1,
-		ok: { success: false, command: "render", error: `Prompt "${parentName}" not found` },
-	};
-}
-
-function rendered(name: string, sections: Record<string, string>, version = 1) {
-	return {
-		success: true,
-		command: "render",
-		name,
-		version,
-		sections: Object.entries(sections).map(([n, body]) => ({ name: n, body })),
-	};
-}
-
-const FAKE_CLONE = async () => ({ cloned: false, localDir: CFG.localDir });
+import { buildSpawn, CanopyClient, CFG, FAKE_CLONE, rendered } from "./refresh.test-helpers.ts";
+import { refreshAgentRegistry } from "./refresh.ts";
 
 describe("refreshAgentRegistry", () => {
 	let db: WarrenDb;
@@ -327,715 +204,5 @@ describe("refreshAgentRegistry", () => {
 
 		expect(result.removed).toEqual([]);
 		expect(await agents.get("stale-bot")).not.toBeNull();
-	});
-});
-
-describe("refreshProjectAgents", () => {
-	let db: WarrenDb;
-	let agents: AgentsRepo;
-	let projects: ProjectsRepo;
-
-	beforeEach(async () => {
-		db = await openDatabase({ path: ":memory:" });
-		const adapter = DrizzleAdapter.for(db);
-		agents = new AgentsRepo(adapter);
-		projects = new ProjectsRepo(adapter);
-	});
-
-	afterEach(async () => {
-		await db.close();
-	});
-
-	const seedProject = async (
-		gitUrl = "https://github.com/x/y.git",
-		localPath = "/data/projects/x/y",
-	) => {
-		const p = await projects.create({ gitUrl, localPath, defaultBranch: "main" });
-		return p.id;
-	};
-
-	test("renders project agents, stamps source=project:<id>, and scopes upserts to the project tier", async () => {
-		const projectId = await seedProject();
-		const spawn = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [
-					{ name: "refactor-bot", version: 1, status: "active" },
-					{ name: "docs-bot", version: 2, status: "active" },
-				],
-			},
-			{
-				"refactor-bot": { ok: rendered("refactor-bot", { system: "proj" }) },
-				"docs-bot": { ok: rendered("docs-bot", { system: "docs" }, 2) },
-			},
-		);
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-		const result = await refreshProjectAgents({ client, agents, projectId });
-
-		expect(result.projectId).toBe(projectId);
-		expect(result.registered.map((r) => r.name).sort()).toEqual(["docs-bot", "refactor-bot"]);
-		expect(result.registered.every((r) => r.projectId === projectId)).toBe(true);
-		expect(result.skipped).toEqual([]);
-		expect(result.removed).toEqual([]);
-
-		const row = await agents.require("refactor-bot", { projectId });
-		const stamped = row.renderedJson as { frontmatter: { source: string } };
-		expect(stamped.frontmatter.source).toBe(`project:${projectId}`);
-		// Global tier untouched.
-		expect(await agents.get("refactor-bot")).toBeNull();
-	});
-
-	test("does not touch a same-named global-tier row", async () => {
-		const projectId = await seedProject();
-		await agents.upsert({
-			name: "claude-code",
-			renderedJson: { frontmatter: { source: "builtin" }, tier: "global" },
-		});
-		const spawn = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [{ name: "claude-code", version: 1, status: "active" }],
-			},
-			{ "claude-code": { ok: rendered("claude-code", { system: "proj override" }) } },
-		);
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-		await refreshProjectAgents({ client, agents, projectId });
-
-		const global = await agents.require("claude-code");
-		expect(global.projectId).toBeNull();
-		expect((global.renderedJson as { tier?: string }).tier).toBe("global");
-		const project = await agents.require("claude-code", { projectId });
-		expect(project.projectId).toBe(projectId);
-		expect((project.renderedJson as { frontmatter: { source: string } }).frontmatter.source).toBe(
-			`project:${projectId}`,
-		);
-	});
-
-	test("skips a render-time failure without aborting the rest", async () => {
-		const projectId = await seedProject();
-		const spawn = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [
-					{ name: "good-bot", version: 1, status: "active" },
-					{ name: "raced-bot", version: 1, status: "active" },
-				],
-			},
-			{
-				"good-bot": { ok: rendered("good-bot", { system: "ok" }) },
-				"raced-bot": {
-					exit: 1,
-					ok: { success: false, command: "render", error: 'Prompt "raced-bot" not found' },
-				},
-			},
-		);
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-		const result = await refreshProjectAgents({ client, agents, projectId });
-		expect(result.registered.map((r) => r.name)).toEqual(["good-bot"]);
-		expect(result.skipped).toHaveLength(1);
-		expect(result.skipped[0]).toMatchObject({
-			name: "raced-bot",
-			code: "canopy_unavailable",
-			reason: expect.stringContaining("not found"),
-		});
-	});
-
-	test("skips a prompt that fails warren's semantic schema (missing system)", async () => {
-		const projectId = await seedProject();
-		const spawn = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [
-					{ name: "bad-bot", version: 1, status: "active" },
-					{ name: "ok-bot", version: 1, status: "active" },
-				],
-			},
-			{
-				"bad-bot": { ok: rendered("bad-bot", { skills: "no system" }) },
-				"ok-bot": { ok: rendered("ok-bot", { system: "ok" }) },
-			},
-		);
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-		const result = await refreshProjectAgents({ client, agents, projectId });
-		expect(result.registered.map((r) => r.name)).toEqual(["ok-bot"]);
-		expect(result.skipped[0]).toMatchObject({
-			name: "bad-bot",
-			code: "agent_schema_error",
-			reason: expect.stringContaining("system"),
-		});
-	});
-
-	test("re-running upserts (lastRefreshed bumps, registeredAt preserved)", async () => {
-		const projectId = await seedProject();
-		const spawnV1 = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [{ name: "refactor-bot", version: 1, status: "active" }],
-			},
-			{ "refactor-bot": { ok: rendered("refactor-bot", { system: "v1" }) } },
-		);
-		await refreshProjectAgents({
-			client: CanopyClient.forProjectPath({ projectPath: "/proj", spawn: spawnV1 }),
-			agents,
-			projectId,
-			now: () => new Date("2026-05-08T12:00:00.000Z"),
-		});
-
-		const spawnV2 = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [{ name: "refactor-bot", version: 2, status: "active" }],
-			},
-			{ "refactor-bot": { ok: rendered("refactor-bot", { system: "v2" }, 2) } },
-		);
-		await refreshProjectAgents({
-			client: CanopyClient.forProjectPath({ projectPath: "/proj", spawn: spawnV2 }),
-			agents,
-			projectId,
-			now: () => new Date("2026-05-09T12:00:00.000Z"),
-		});
-
-		const row = await agents.require("refactor-bot", { projectId });
-		expect(row.registeredAt).toBe("2026-05-08T12:00:00.000Z");
-		expect(row.lastRefreshed).toBe("2026-05-09T12:00:00.000Z");
-		expect((row.renderedJson as { version: number }).version).toBe(2);
-	});
-
-	test("prune is always-on: rows missing from this project's listing are deleted; global + other-project rows untouched", async () => {
-		const projectId = await seedProject();
-		const otherProjectId = await seedProject("https://github.com/a/b.git", "/data/projects/a/b");
-		// Pre-existing rows across all three scopes:
-		await agents.upsert({ name: "stale-bot", projectId, renderedJson: { v: 0 } });
-		await agents.upsert({ name: "stale-bot", renderedJson: { tier: "global" } });
-		await agents.upsert({ name: "stale-bot", projectId: otherProjectId, renderedJson: { v: 0 } });
-
-		const spawn = buildSpawn(
-			{
-				success: true,
-				command: "list",
-				prompts: [{ name: "live-bot", version: 1, status: "active" }],
-			},
-			{ "live-bot": { ok: rendered("live-bot", { system: "ok" }) } },
-		);
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-		const result = await refreshProjectAgents({ client, agents, projectId });
-		expect(result.removed).toEqual(["stale-bot"]);
-		expect(await agents.get("stale-bot", { projectId })).toBeNull();
-		// Global tier untouched.
-		expect((await agents.get("stale-bot"))?.projectId).toBeNull();
-		// Other project untouched.
-		expect((await agents.get("stale-bot", { projectId: otherProjectId }))?.projectId).toBe(
-			otherProjectId,
-		);
-	});
-
-	test("empty listing with no existing rows is a no-op", async () => {
-		const projectId = await seedProject();
-		const spawn = buildSpawn({ success: true, command: "list", prompts: [] }, {});
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-		const result = await refreshProjectAgents({ client, agents, projectId });
-		expect(result).toEqual({ projectId, registered: [], skipped: [], removed: [] });
-	});
-
-	test("transport-layer failure on `cn list` aborts the whole refresh", async () => {
-		const projectId = await seedProject();
-		const spawn: SpawnFn = async () => ({
-			stdout: "",
-			stderr: "cn: command not found",
-			exitCode: 127,
-		});
-		const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-		await expect(refreshProjectAgents({ client, agents, projectId })).rejects.toMatchObject({
-			code: "canopy_unavailable",
-		});
-	});
-
-	describe("cross-tier inheritance (warren-44a3)", () => {
-		test("composes a project-tier role that extends a built-in agent", async () => {
-			const projectId = await seedProject();
-			// Pre-seed the global tier with a "claude-code" built-in.
-			await agents.upsert({
-				name: "claude-code",
-				renderedJson: {
-					name: "claude-code",
-					version: 1,
-					sections: { system: "builtin-system", workflow: "builtin-workflow" },
-					resolvedFrom: ["claude-code"],
-					frontmatter: { source: "builtin", provider: "anthropic" },
-				},
-			});
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "refactor-bot", version: 1, status: "active" }],
-				},
-				{
-					// cn render fails because canopy can't find the cross-tier parent.
-					"refactor-bot": renderMissingParent("claude-code"),
-				},
-				{
-					// cn show returns the raw project prompt with extends: claude-code.
-					"refactor-bot": showOk("refactor-bot", {
-						sections: { system: "refactor-system", expertise_seed: "refactor-seed" },
-						extends: "claude-code",
-						frontmatter: { model: "claude-sonnet-4-6" },
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-
-			expect(result.skipped).toEqual([]);
-			expect(result.registered.map((r) => r.name)).toEqual(["refactor-bot"]);
-			const row = await agents.require("refactor-bot", { projectId });
-			const def = row.renderedJson as {
-				sections: Record<string, string>;
-				frontmatter: { source: string; provider?: string; model?: string };
-				resolvedFrom: string[];
-			};
-			// Built-in system overridden, built-in workflow preserved, project's expertise_seed added.
-			expect(def.sections).toEqual({
-				system: "refactor-system",
-				workflow: "builtin-workflow",
-				expertise_seed: "refactor-seed",
-			});
-			// Source stamp reflects the LEAF tier, not the built-in parent.
-			expect(def.frontmatter.source).toBe(`project:${projectId}`);
-			// Parent frontmatter merged in; focal frontmatter still applied; source overlay last.
-			expect(def.frontmatter.provider).toBe("anthropic");
-			expect(def.frontmatter.model).toBe("claude-sonnet-4-6");
-			expect(def.resolvedFrom).toEqual(["claude-code", "refactor-bot"]);
-		});
-
-		test("composes a project-tier role whose chain walks through a library parent", async () => {
-			const projectId = await seedProject();
-			// Library-tier parent (no source stamp; readAgentSource collapses to library).
-			await agents.upsert({
-				name: "library-base",
-				renderedJson: {
-					name: "library-base",
-					version: 2,
-					sections: { system: "lib-system", verbose: "lib-verbose" },
-					resolvedFrom: ["library-base"],
-					frontmatter: {},
-				},
-			});
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "tuned", version: 1, status: "active" }],
-				},
-				{ tuned: renderMissingParent("library-base") },
-				{
-					tuned: showOk("tuned", {
-						sections: { skills: "tuned-skills" },
-						extends: "library-base",
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-			expect(result.skipped).toEqual([]);
-			const row = await agents.require("tuned", { projectId });
-			const def = row.renderedJson as {
-				sections: Record<string, string>;
-				frontmatter: { source: string };
-			};
-			expect(def.sections).toEqual({
-				system: "lib-system",
-				verbose: "lib-verbose",
-				skills: "tuned-skills",
-			});
-			expect(def.frontmatter.source).toBe(`project:${projectId}`);
-		});
-
-		test("source stamping walks past a same-named project-tier shadow to a real built-in parent", async () => {
-			// The seed's open question: project has its own role NAMED claude-code
-			// AND another role that `extends: claude-code`. Parent resolution
-			// must NOT loop on the project shadow — it should bottom out at the
-			// global built-in. Because both project prompts live in the project's
-			// .canopy/, canopy's own resolver actually handles this case (in-tier
-			// extends). We assert the merge result is correct end-to-end.
-			const projectId = await seedProject();
-			await agents.upsert({
-				name: "claude-code",
-				renderedJson: {
-					name: "claude-code",
-					version: 1,
-					sections: { system: "builtin", workflow: "wf" },
-					resolvedFrom: ["claude-code"],
-					frontmatter: { source: "builtin" },
-				},
-			});
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [
-						// The project-tier shadow of claude-code itself extends the built-in.
-						{ name: "claude-code", version: 1, status: "active" },
-						{ name: "tuned", version: 1, status: "active" },
-					],
-				},
-				{
-					// Both renders bail because the chain points outside the project.
-					"claude-code": renderMissingParent("claude-code"),
-					tuned: renderMissingParent("claude-code"),
-				},
-				{
-					"claude-code": showOk("claude-code", {
-						sections: { workflow: "project-workflow" },
-						extends: "claude-code",
-					}),
-					tuned: showOk("tuned", {
-						sections: { system: "tuned-system" },
-						extends: "claude-code",
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-			expect(result.skipped).toEqual([]);
-			const tuned = (await agents.require("tuned", { projectId })).renderedJson as {
-				sections: Record<string, string>;
-				frontmatter: { source: string };
-				resolvedFrom: string[];
-			};
-			// tuned → project-tier "claude-code" (workflow override) → built-in "claude-code"
-			// (system + workflow). tuned's own system wins.
-			expect(tuned.sections).toEqual({
-				system: "tuned-system",
-				workflow: "project-workflow",
-			});
-			expect(tuned.frontmatter.source).toBe(`project:${projectId}`);
-			// Built-in claude-code, project shadow, focal — three hops.
-			expect(tuned.resolvedFrom).toEqual(["claude-code", "claude-code", "tuned"]);
-		});
-
-		test("when compose fails because the parent doesn't exist anywhere, surface the compose error in skipped", async () => {
-			const projectId = await seedProject();
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "orphan", version: 1, status: "active" }],
-				},
-				{ orphan: renderMissingParent("missing-parent") },
-				{
-					orphan: showOk("orphan", {
-						sections: { system: "orphan-system" },
-						extends: "missing-parent",
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-			expect(result.registered).toEqual([]);
-			expect(result.skipped[0]).toMatchObject({
-				name: "orphan",
-				code: "agent_schema_error",
-				reason: expect.stringContaining("missing-parent"),
-			});
-		});
-
-		test("falls back to the original canopy skip when the focal prompt has no inheritance", async () => {
-			// A "Prompt not found" failure on a focal prompt with NO extends/mixins
-			// is a genuine missing-prompt case (e.g. archived between list and
-			// render) — compose should not mask it, and the original canopy
-			// error should be returned.
-			const projectId = await seedProject();
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "raced-bot", version: 1, status: "active" }],
-				},
-				{
-					"raced-bot": {
-						exit: 1,
-						ok: { success: false, command: "render", error: `Prompt "raced-bot" not found` },
-					},
-				},
-				{
-					"raced-bot": showOk("raced-bot", {
-						sections: { system: "x" },
-						// No extends, no mixins → compose returns null and the canopy
-						// render skip survives.
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-			expect(result.registered).toEqual([]);
-			expect(result.skipped[0]).toMatchObject({
-				name: "raced-bot",
-				code: "canopy_unavailable",
-				reason: expect.stringContaining("raced-bot"),
-			});
-		});
-
-		test("composed agent missing required system section is surfaced as agent_schema_error", async () => {
-			const projectId = await seedProject();
-			await agents.upsert({
-				name: "minimal-base",
-				renderedJson: {
-					name: "minimal-base",
-					version: 1,
-					sections: { skills: "lib-skills" }, // no system section anywhere in the chain
-					resolvedFrom: ["minimal-base"],
-					frontmatter: {},
-				},
-			});
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "incomplete", version: 1, status: "active" }],
-				},
-				{ incomplete: renderMissingParent("minimal-base") },
-				{
-					incomplete: showOk("incomplete", {
-						sections: { extras: "x" },
-						extends: "minimal-base",
-					}),
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-
-			const result = await refreshProjectAgents({ client, agents, projectId });
-			expect(result.registered).toEqual([]);
-			expect(result.skipped[0]).toMatchObject({
-				name: "incomplete",
-				code: "agent_schema_error",
-				reason: expect.stringContaining("system"),
-			});
-		});
-	});
-
-	describe("on-disk rendered cache (warren-44e3)", () => {
-		const trackWriter = (): {
-			writer: RenderedCacheWriter;
-			calls: {
-				init: string[];
-				write: Array<{ projectPath: string; name: string; definition: AgentDefinition }>;
-				prune: Array<{ projectPath: string; name: string }>;
-			};
-		} => {
-			const calls = {
-				init: [] as string[],
-				write: [] as Array<{
-					projectPath: string;
-					name: string;
-					definition: AgentDefinition;
-				}>,
-				prune: [] as Array<{ projectPath: string; name: string }>,
-			};
-			const writer: RenderedCacheWriter = {
-				async init(projectPath) {
-					calls.init.push(projectPath);
-				},
-				async write(projectPath, name, definition) {
-					calls.write.push({ projectPath, name, definition });
-				},
-				async prune(projectPath, name) {
-					calls.prune.push({ projectPath, name });
-				},
-			};
-			return { writer, calls };
-		};
-
-		test("does not invoke the cache writer when projectPath is omitted", async () => {
-			const projectId = await seedProject();
-			const { writer, calls } = trackWriter();
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [{ name: "docs-bot", version: 1, status: "active" }],
-				},
-				{ "docs-bot": { ok: rendered("docs-bot", { system: "ok" }) } },
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-			await refreshProjectAgents({ client, agents, projectId, cacheWriter: writer });
-			expect(calls.init).toEqual([]);
-			expect(calls.write).toEqual([]);
-			expect(calls.prune).toEqual([]);
-		});
-
-		test("calls init once, write per registered agent, prune per removed row", async () => {
-			const projectId = await seedProject();
-			// Pre-seed a stale row that will be pruned on this refresh.
-			await agents.upsert({ name: "stale-bot", projectId, renderedJson: { v: 0 } });
-			const { writer, calls } = trackWriter();
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [
-						{ name: "refactor-bot", version: 1, status: "active" },
-						{ name: "docs-bot", version: 2, status: "active" },
-					],
-				},
-				{
-					"refactor-bot": { ok: rendered("refactor-bot", { system: "rf" }) },
-					"docs-bot": { ok: rendered("docs-bot", { system: "doc" }, 2) },
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-			await refreshProjectAgents({
-				client,
-				agents,
-				projectId,
-				projectPath: "/proj/tree",
-				cacheWriter: writer,
-			});
-			expect(calls.init).toEqual(["/proj/tree"]);
-			expect(calls.write.map((c) => c.name).sort()).toEqual(["docs-bot", "refactor-bot"]);
-			expect(calls.write.every((c) => c.projectPath === "/proj/tree")).toBe(true);
-			// Stamped frontmatter.source survives the write path.
-			const refactor = calls.write.find((c) => c.name === "refactor-bot");
-			expect(refactor?.definition.frontmatter.source).toBe(`project:${projectId}`);
-			expect(calls.prune).toEqual([{ projectPath: "/proj/tree", name: "stale-bot" }]);
-		});
-
-		test("skipped agents (render or schema failure) do not produce a cache write", async () => {
-			const projectId = await seedProject();
-			const { writer, calls } = trackWriter();
-			const spawn = buildSpawn(
-				{
-					success: true,
-					command: "list",
-					prompts: [
-						{ name: "ok-bot", version: 1, status: "active" },
-						{ name: "bad-bot", version: 1, status: "active" },
-					],
-				},
-				{
-					"ok-bot": { ok: rendered("ok-bot", { system: "ok" }) },
-					"bad-bot": { ok: rendered("bad-bot", { skills: "no system" }) },
-				},
-			);
-			const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-			await refreshProjectAgents({
-				client,
-				agents,
-				projectId,
-				projectPath: "/proj/tree",
-				cacheWriter: writer,
-			});
-			expect(calls.write.map((c) => c.name)).toEqual(["ok-bot"]);
-		});
-
-		test("default writer writes <projectPath>/.canopy/.rendered/<name>.json and seeds .gitignore", async () => {
-			const projectId = await seedProject();
-			const projectPath = await mkdtemp(join(tmpdir(), "warren-44e3-"));
-			try {
-				const spawn = buildSpawn(
-					{
-						success: true,
-						command: "list",
-						prompts: [{ name: "docs-bot", version: 2, status: "active" }],
-					},
-					{ "docs-bot": { ok: rendered("docs-bot", { system: "doc body" }, 2) } },
-				);
-				const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-				await refreshProjectAgents({ client, agents, projectId, projectPath });
-
-				const cacheDir = join(projectPath, RENDERED_CACHE_SUBPATH);
-				const entries = (await readdir(cacheDir)).sort();
-				expect(entries).toEqual([".gitignore", "docs-bot.json"]);
-				const gitignore = await readFile(join(cacheDir, ".gitignore"), "utf8");
-				expect(gitignore).toBe("*\n");
-				const cached = JSON.parse(await readFile(join(cacheDir, "docs-bot.json"), "utf8")) as {
-					name: string;
-					version: number;
-					sections: Record<string, string>;
-					frontmatter: { source: string };
-				};
-				expect(cached.name).toBe("docs-bot");
-				expect(cached.version).toBe(2);
-				expect(cached.sections.system).toBe("doc body");
-				expect(cached.frontmatter.source).toBe(`project:${projectId}`);
-			} finally {
-				await rm(projectPath, { recursive: true, force: true });
-			}
-		});
-
-		test("default writer prunes the JSON when a row is removed", async () => {
-			const projectId = await seedProject();
-			const projectPath = await mkdtemp(join(tmpdir(), "warren-44e3-"));
-			try {
-				// First refresh registers stale-bot AND writes its cache file.
-				const spawn1 = buildSpawn(
-					{
-						success: true,
-						command: "list",
-						prompts: [{ name: "stale-bot", version: 1, status: "active" }],
-					},
-					{ "stale-bot": { ok: rendered("stale-bot", { system: "x" }) } },
-				);
-				await refreshProjectAgents({
-					client: CanopyClient.forProjectPath({ projectPath: "/proj", spawn: spawn1 }),
-					agents,
-					projectId,
-					projectPath,
-				});
-				const cacheDir = join(projectPath, RENDERED_CACHE_SUBPATH);
-				expect((await readdir(cacheDir)).includes("stale-bot.json")).toBe(true);
-
-				// Second refresh drops stale-bot from the listing → file removed.
-				const spawn2 = buildSpawn({ success: true, command: "list", prompts: [] }, {});
-				const result = await refreshProjectAgents({
-					client: CanopyClient.forProjectPath({ projectPath: "/proj", spawn: spawn2 }),
-					agents,
-					projectId,
-					projectPath,
-				});
-				expect(result.removed).toEqual(["stale-bot"]);
-				expect((await readdir(cacheDir)).includes("stale-bot.json")).toBe(false);
-			} finally {
-				await rm(projectPath, { recursive: true, force: true });
-			}
-		});
-
-		test("default writer skips unsafe agent names (defense-in-depth at the filesystem boundary)", async () => {
-			const projectId = await seedProject();
-			const projectPath = await mkdtemp(join(tmpdir(), "warren-44e3-"));
-			try {
-				const spawn = buildSpawn(
-					{
-						success: true,
-						command: "list",
-						prompts: [{ name: "../escape", version: 1, status: "active" }],
-					},
-					{ "../escape": { ok: rendered("../escape", { system: "x" }) } },
-				);
-				const client = CanopyClient.forProjectPath({ projectPath: "/proj", spawn });
-				await refreshProjectAgents({ client, agents, projectId, projectPath });
-				const cacheDir = join(projectPath, RENDERED_CACHE_SUBPATH);
-				// Only the .gitignore was seeded; no JSON written for the unsafe name.
-				expect(await readdir(cacheDir)).toEqual([".gitignore"]);
-			} finally {
-				await rm(projectPath, { recursive: true, force: true });
-			}
-		});
 	});
 });
